refactor: rename touch header
